Split up commands into files
Purely cosmetic so far, not a single functionality has been tested.
