+++ commands/general/blog.js
@@ -1,0 +1,7 @@
+exports.run = (arg, by, room) => {
+    this.say(room, "/addhtmlbox <a href='https://spo.ink/ansena'>ansena's blog</a>");
+};
+
+module.exports.config = {
+    aliases: []
+};
