+++ commands/general/chef.js
@@ -1,0 +1,7 @@
+exports.run = (arg, by, room) => {
+    this.say(room, "!dt sheer cold");
+};
+
+module.exports.config = {
+    aliases: []
+};
