+++ commands/general/conics.js
@@ -1,0 +1,7 @@
+exports.run = (arg, by, room) => {
+    this.say(room, "!dt mudkip");
+};
+
+module.exports.config = {
+    aliases: []
+};
