+++ commands/general/delet.js
@@ -1,0 +1,7 @@
+exports.run = (arg, by, room) => {
+    this.say(room, arg + " **deleted**.");
+};
+
+module.exports.config = {
+    aliases: []
+};
