+++ commands/general/epic.js
@@ -1,0 +1,7 @@
+exports.run = (arg, by, room) => {
+    this.say(room, "gaming");
+};
+
+module.exports.config = {
+    aliases: []
+};
