+++ commands/general/nom.js
@@ -1,0 +1,7 @@
+exports.run = (arg, by, room) => {
+    this.say(room, "Player not recognized. Perhaps you meant **seaco**.");
+};
+
+module.exports.config = {
+    aliases: []
+};
