+++ commands/general/vgc.js
@@ -1,0 +1,8 @@
+exports.run = (arg, by, room) => {
+    let text = 'indeed \ud83c\udd71\ufe0frother, vgc';
+    this.say(room, text);
+};
+
+module.exports.config = {
+    aliases: []
+};
